[Client Refactoring 8] Add AuthorityAggregator (#364)
* [Client Refactoring] Add AuthorityAggregator

* [Client Refactoring 9] Make download_objects_not_in_db fast again (#368)

* [Client Refactoring 9] Make download_objects_not_in_db fast again

* [Clienet Refactoring 10] Properly lock orde forr transfer_to_fastx_unsafe_unconfirmed (#369)
