+++ fastpay_core/src/lib.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod authority;
+pub mod authority_aggregator;
 pub mod authority_client;
 pub mod authority_server;
 pub mod client;
